+++ src/parser/CellAddress.ts
@@ -5,6 +5,7 @@
 
 import {
   absoluteSheetReference,
+  invalidSimpleCellAddress,
   simpleCellAddress,
   SimpleCellAddress,
   simpleColumnAddress,
@@ -12,8 +13,9 @@ import {
   simpleRowAddress,
   SimpleRowAddress,
 } from '../Cell'
-import {columnIndexToLabel} from './addressRepresentationConverters'
+import {Maybe} from '../Maybe'
 import {AddressWithColumn, AddressWithRow} from './Address'
+import {columnIndexToLabel} from './addressRepresentationConverters'
 
 /** Possible kinds of cell references */
 export enum CellReferenceType {
@@ -158,8 +160,11 @@ export class CellAddress implements AddressWithColumn, AddressWithRow {
     }
   }
 
-  public unparse(baseAddress: SimpleCellAddress): string {
+  public unparse(baseAddress: SimpleCellAddress): Maybe<string> {
     const simpleAddress = this.toSimpleCellAddress(baseAddress)
+    if(invalidSimpleCellAddress(simpleAddress)) {
+      return undefined
+    }
     const column = columnIndexToLabel(simpleAddress.col)
     const rowDollar = this.type === CellReferenceType.CELL_REFERENCE_ABSOLUTE || this.type === CellReferenceType.CELL_REFERENCE_ABSOLUTE_ROW ? '$' : ''
     const colDollar = this.type === CellReferenceType.CELL_REFERENCE_ABSOLUTE || this.type === CellReferenceType.CELL_REFERENCE_ABSOLUTE_COL ? '$' : ''